requirements.txt
python-telegram-bot==20.3
apscheduler==3.10.1
pytz==2023.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0